+++ includes/Db/UserMapper.php
@@ -35,11 +35,12 @@ class UserMapper extends Mapper
     public function save(User $user): bool
     {
         if (empty($user->getUid())) {
-            $sql = 'INSERT INTO user (active, username, hash, email, honorific, name_first, ';
-            $sql .= 'name_last, company, website, address_street, address_suburb, address_city, address_state, ';
-            $sql .= 'address_country, address_postcode, phone_mobile, phone_work, password_reset, password_reset_ttl)';
+            $sql = 'INSERT INTO user (active, username, hash, email, honorific, name_first,';
+            $sql .= ' name_last, company, website, address_street, address_suburb, address_city, address_state,';
+            // phpcs:ignore
+            $sql .= ' address_country, address_postcode, phone_mobile, phone_work, password_reset, password_reset_ttl, refresh_token)';
             $sql .= ' VALUES';
-            $sql .= ' (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)';
+            $sql .= ' (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)';
             $bindParams = [
                 $user->getActive(),
                 $user->getUsername(),
@@ -60,12 +61,13 @@ class UserMapper extends Mapper
                 $user->getPhoneWork(),
                 $user->getPasswordReset(),
                 $user->getPasswordResetTtl(),
+                $user->getRefreshToken(),
             ];
         } else {
             $sql = 'UPDATE user SET active=?, username=?, hash=?, email=?, honorific=?, ';
             $sql .= 'name_first=?, name_last=?, company=?, website=?, address_street=?, address_suburb=?, ';
             $sql .= 'address_city=?, address_state=?, address_country=?, address_postcode=?, phone_mobile=?, ';
-            $sql .= 'phone_work=?, password_reset=?, password_reset_ttl=? WHERE uid=?';
+            $sql .= 'phone_work=?, password_reset=?, password_reset_ttl=?, refresh_token=? WHERE uid=?';
             $bindParams = [
                 $user->getActive(),
                 $user->getUsername(),
@@ -86,6 +88,7 @@ class UserMapper extends Mapper
                 $user->getPhoneWork(),
                 $user->getPasswordReset(),
                 $user->getPasswordResetTtl(),
+                $user->getRefreshToken(),
                 $user->getUid(),
             ];
         }
@@ -144,7 +147,7 @@ class UserMapper extends Mapper
     }
 
     /**
-     * Find allUsers
+     * Find all users
      *
      * @return array User objects.
      *
@@ -234,6 +237,7 @@ class UserMapper extends Mapper
         $user->setActive($row['active'] ?? 0);
         $user->setUsername($row['username'] ?? '');
         $user->setHash($row['hash'] ?? null);
+        $user->setRefreshToken($row['refresh_token'] ?? null);
         $user->setEmail($row['email'] ?? '');
         $user->setHonorific($row['honorific'] ?? null);
         $user->setNameFirst($row['name_first'] ?? null);