+++ includes/Processor/ApplicationDelete.php
@@ -108,7 +108,7 @@ class ApplicationDelete extends ProcessorEntity
         parent::process();
 
         try {
-            $uid = Utilities::getUidFromToken();
+            $uid = Utilities::getClaimFromToken('uid');
         } catch (ApiException $e) {
             throw new ApiException($e->getMessage(), $e->getCode(), $this->id, $e->getHtmlCode());
         }