+++ includes/Processor/OpenapiUpdate.php
@@ -132,7 +132,7 @@ class OpenapiUpdate extends ProcessorEntity
         }
         try {
             $application = $this->applicationMapper->findByAccidAppname($account->getAccid(), $applicationName);
-            $roles = Utilities::getRolesFromToken();
+            $roles = Utilities::getClaimFromToken('roles');
         } catch (ApiException $e) {
             throw new ApiException($e->getMessage(), $e->getCode(), $this->id, $e->getHtmlCode());
         }