+++ includes/Processor/ResourceCreate.php
@@ -302,7 +302,7 @@ class ResourceCreate extends ProcessorEntity
 
         // Validate user has developer role for the application
         try {
-            $userRoles = Utilities::getRolesFromToken();
+            $userRoles = Utilities::getClaimFromToken('roles');
         } catch (ApiException $e) {
             throw new ApiException($e->getMessage(), $e->getCode(), $this->id, $e->getHtmlCode());
         }